premiers exercices en C

+++ pointers_arrays_strings/2-strlen.c
@@ -10,9 +10,9 @@
  */
 
 
-long long int _strlen(char *s)
+long int _strlen(char *s)
 {
-long long int len = 0;
+long int len = 0;
 len = sizeof(s) + 2;
 return len;
 }
